requirements.txt
PyQt5==5.15.7
ffmpeg-python==0.2.0
requests==2.31.0
Pillow==9.5.0
pathlib2==2.3.7
pytest==7.4.3 